WARNING: *** THIS COMMIT WILL BREAK THE SCRIPTER ***
Take a deep breath before reading this... The latest commit of the 
library is an all-scale re-implementation of PHOEBE parameters. The 
partial commit is done so that Gal and Olivera can build the widget 
table with the up-to-specs code instead of the old obsolete constructs. 
Yet be sure to take extra care, because numerous changes have hit the 
SVN and some of the functions (i.e. minimizers, open/save error 
handling etc) are known not to work. Do not update to this version if 
you want a functional scripter! You have been warned. ;) But now that 
you've done it anyway, please wait until tomorrow, I am working on it 
around the clock.