Initial gtk-doc support added to the library. All build files have been updated, a new directory docs/reference has been created. gtk-doc will not be used by default; if you want to use it, pass --enable-gtk-doc switch to ./configure.
The basic setup seems to work fine, but gtk-doc segfaults on the current 
setup. It probably has something to do with the formatting rules for the 
comments. Hopefully Olivera can look into this. Another issue is the 
addition of GTK+ CFLAGS and LIBS in docs/reference/Makefile.am - it is 
not final yet, but it seems to work for now.