gtk-doc is failing on the automatically created lex and tab sources; these header files have been added to the list of ignored headers in docs/reference/Makefile.am; Makefile.in has been regenerated.
On a more important side, infrastructure for PHOEBE_parameter_table 
structures has been set up. Since the API should not change, another 
global variable PHOEBE_pt_list was introduced that holds all defined 
parameter tables. The active table is selected by the
phoebe_parameter_table_activate () function; this sets a global pointer 
PHOEBE_pt to point to the selected table in the pool of tables. This way 
all functions for parameter manipulation act on the currently selected 
table.

Next to the infrastructure, table duplication function 
phoebe_parameter_table_duplicate () has been written; it has to be 
performed in two stages: the first stage copies the contents of 
parameters, and the second stage copies the dependencies. The traversal 
needs to be done in stages because all parameters need to be defined to 
be able to set the dependencies. The second stage still needs to be 
done.

The next step is to implement table duplication for the minimizers.