1. Expanded the treeview model for spots, so that it now contains adjustment information for all spot parameters.
2. Turned the load-lc, load-rv and load-spots windows to GtkDialogs. 
Now all the widgets inside these windows are declared and initialized 
upon opening the dialog, and destroyed with the dialog after it returns 
control to the main window. If "OK" is pressed, the widgets are read 
out, and their values are assigned to a new row of the corresponding 
treeview model. Parameters that aren't entered trough these  dialogs
are still hard-coded to their default values.

3. Deleted some obsolete signals from phoebe.glade, and removed the lot 
of signal handlers for the above-mentioned dialogs from 
phoebe_gui_callbakcs.