The "edit" buttons for data->lc, data->rv and spots treeviews now work. Only elementary error handling is in place, so some warnings may appear when doing strange things (like adding a curve with no filename).
The other thing is more subtle, and open for discussion. So far the 
column types for LCs and RVs and the source of the spots were kept as 
strings in the models. However, they appear as combobox widgets in the 
GUI, and combos reffer to their items by integer indices. To get around
this, I changed the models to contain integers instead, and added a 
series of "cell data functions" that render these indices as corresponding 
strings in the views. However, this might proove to be a performance 
issue, as these functions are called quite often by the treeviews... I 
haven't noticed any impact on my (quite low-end) machine, but I thought 
it was worth mentioning anyway.