Added a block to gui_get_value_from_widget to handle the TreeViewColumn "widgets". To do this, I had to decorate every column with a set of additional data fields, one being a pointer to its parent TreeView, and the other its column number (these are the changes to phoebe_gui_treeviews.c).
To try and test this a bit, I added another toolbutton to the toolbar of 
the main window, which calls gui_get_values_from_widgets. Right now this 
call causes a segfault; by commenting pieces of 
gui_get_value_from_widget I found that the problem is inside the 
CheckButton and SpinButton blocks, but I don't know what the problem is.