Rewritten the TreeViewColumn block of the gui_get_value_from_widget function completly. I'm ashamed to say that I forgot the golden rule and fiddled with the parameters directly, rather than through library functions. I fixed it all now, and as a result, this section doesn't break Phoebe any more (even after the latest fixes, it stil used to segfault if any curves were added). One strange problem remains: after the "Update parameters" is pressed, both LC and RV treevies in the Data tab start behaving strangely...
I also made the phoebe_gui_types.c a little more tidy - aligned the code 
a bit and moved some parts around. There are so many lines of similar 
code there that it was barely readble. I know this is silly, but please 
don't discard these changes if they cause conflicts (and they are likely 
to) - they took me a considerable amount of time. Thank you :)