Added a new tab for spots. It works like this: click a button to add a spot to the primary or secondary, and then use the spinbuttons below to change the parameters for the selected spot.
The sistem isn't quite finished yet: the connection to the library 
parameters is yet to be implemented, and for some reason the new spots 
appear as "tba" when they are added to the treeview; also, the spot 
labels aren't computed properly, add several spots and watch the labels 
below to see this effect. I'll get back to this tomorrow.

There are also some problems in the glade file because I tried to shrink 
the main window. Will fix later.

Until then, try out and comment! Andrej, if you want to change something 
about the spots in the library, now is a good time to make a review...