I made many changes, sorry to put them all in one commit.
The most important is the handler for the "Calculate" button on the 
fitting tab, that now calls DC minimizer. To make this work, I also had 
to change the way the filters are stored in LC an RV treeviews, and 
while at that, I moved the add and remove handlers for those treeviews 
to phoebe_gui_treeviews.c. Then I went on to erase all the empty signal 
handlers from phoebe_gui_callbacks.c, and to remove them from the glade 
file as well.

I removed the uber "adjustible_spots_model" from phoebe_gui_types.c, but 
left the other spot-related stuff untouched: I still need to think about 
the effects the new library functions will have on the current layout 
and functionality.

Last, but not least, I added a return statement to if(!widget->par) 
block in gui_set/get_values functions, so the "I got where I'm not 
supposed to get!" warnings are gone.