Sigh... After implementing everything in phoebe_spots.c, I realized it is not a good way to go because it deviates from the rest of the code too much. As consequence I ditched the structures and introduced some additional spot parameters instead.
The layout is now the following:

1) Hook up tree-view spot parameters to:

	phoebe_spots_no
	phoebe_spots_active_switch
	phoebe_spots_source
	phoebe_spots_colatitude
	phoebe_spots_longitude
	phoebe_spots_radius
	phoebe_spots_tempfactor

2) Adjustable widgets below the tree-view pertain to:

	wd_spots_lat1
	wd_spots_long1
	wd_spots_rad1
	wd_spots_temp1
	wd_spots_lat2
	wd_spots_long2
	wd_spots_rad2
	wd_spots_temp2

   However, the correspondence is not 1:1 because there can be many 
   spots on either star. The "trick" is to use phoebe_spots_tba_switch, 
   which is defined in phoebe_parameters.c but currently commented out. 
   If you want to use it, just uncomment it. I commented it out because 
   I wanted to give it a bit more thought as to what would be optimal.

The changes to the code to reflect these changes in organization are 
close to complete. It's up to the driving program to assign values to 
wd_spots_* properly, but this might be better in the library. I have to 
think about this a little bit more.

While changing everything, qualifiers phoebe_spots_move1 and 
phoebe_spots_move2 renamed to phoebe_spots_corotate1 and 
phoebe_spots_corotate2. I suggest a similar renaming of the 
corresponding check-buttons in the GUI.