Attached the spots treeview to the new library parameters. The only change in the library itself is the uncommented line defining the tba parameter for a spot. I had to make a lot of small changes to GUI sources to  get this through, and not everything is behaving; if you save a parameter file with some spots, chances are Phoebe will segfault when you later open it.
I also added a flag to show or not to show all the columns in the spots 
treeview, so if we decide the users should be able to expand it, we can 
let them.

In the meantime, I fixed the detach button images, which now display 
properly. Two or three libglade warnings remain, though.