Spot handling fully functional. The following parameters have been defined:
	phoebe_spots_colatitude_tba
	phoebe_spots_longitude_tba
	phoebe_spots_radius_tba
	phoebe_spots_tempfactor_tba

These are all boolean arrays and expect 1 or 0 (TRUE or FALSE) for each 
spot separately. The newly implemented function wd_spots_parameters_get 
() is a wrapper that transfers the data from PHOEBE spots parameters 
phoebe_spots_* to WD spots parameters wd_spots_*. This is done to 
temporarily (until we implement our version of DC) support up to two 
adjustable spots in DC *and/or* NMS.

To recap: if you want to adjust, say, radius of the third spot and 
colatitude on the fourth spot, you would set:

	phoebe_parameter_set_value (phoebe_spots_radius_tba, 1, 3);
	phoebe_parameter_set_value (phoebe_spots_colatitude_tba, 1, 4);

Note that the source star for these spots is defined by the 3rd and 4th 
element of the phoebe_spots_source parameter.