Reading out legacy parameter files implemented. The adopted scheme is:
1) Detect 0.2x keyword file
2) Run a flex-based converter to get a 0.3x compliant parameter file
3) Save that parameter file in a temporary directory
4) Open the parameter file
5) Remove the parameter file

The converter is outside auto* utilities for the time being; if you want 
to recompile it, issue:

  flex -ophoebe_legacy.c --header-file=phoebe_legacy.h phoebe_legacy.lex

This will be autoconfiscated as soon as the converter is stable. The 
conversion already works, but there are some subtle issues that need to 
be still resolved, i.e. spots. See the terminal output after conversion 
to see a list of problems.