Since the data are passed to WD directly, the workarounds for limited formatting have been removed and weights are now consistently assigned and no longer renormalized. These are the transformation rules:
  Undefined          -> Standard deviation: output = sigma
  Undefined          -> Standard weights:   output = 1/sigma^2
  Standard deviation -> Standard weights:   output = 1/input^2
  Standard weights   -> Standard deviation: output = 1/sqrt(input)

Here sigma is the passband standard deviation that is used to obtain 
consistent chi2 values.

PHOEBE still creates dci/dco files for debugging purposes; if weights 
overflow the formatting space (i.e. for values larger than 99.9), they 
are replaced with -1.0. This of course has no effect on the computation 
since the data are passed to DC directly. A caption added to DCO file 
that stresses this (for people that inspect DC outputs manually).