Updated the setup.py so that
python setup.py build
python setup.py install

are now possible. In fact, that should be all you have to do after a fresh
chekout in order to get things working. The setup.py file was moved 
to the sandbox root directory and pyphoebe is installed as a package. 
I tested it with the minimal example and it worked but it will likely 
fail on your end so please report back why and how it failed and I'll 
try to fix it. Note that the setup.py script is very basic right now 
and is missing a lot of standard information.