Initial manager/worker MPI implementation prototype. To see it in action, install mpi4py and run:
    mpirun -np 4 python mpi.py > test.lc

That will generate a test light curve (phase, flux, worker) that you can
plot in, say, gnuplot. None of the actual phoebe backend functions were
used, I only implemented the MPI part to keep it simple. The code is not
optimized and needs to be fool-proofed with try/excepts and asserts. The
number of processors passed to mpirun via the -np switch does not need
to correspond to the actual number of processors; any number may be
passed here, but of course it won't be computationally useful to pass
anything larger than the available number of cores. The code is cluster-
ready, provided that the cluster hostfile is set up correctly. To test
on, say, 24 cores listed in the hosts file, run:

    mpirun -np 24 --hostfile hosts --byslot python mpi.py > test.lc