Porting to devel, and a lot of updates:
(0) changed some parameter names conform the workshop decisions
(1) re-implemented keplerorbits for binary and implemented hierarchical multiple-body orbits
(2) added support for LTT in binary orbits and hieararchical orbits; Roemer delay via Newton-Raphson
(3) changed names of parameters in keplerorbit.py to match the phoebe context names
(4) Ported rotation and translation to C (cgeometry.c)
(5) re-implemented atmosphere table framework to be more flexible: any parameter
can now be interpolated as long as it is provided in the tables of specific
intensities. The Kurucz grids can be interpolated in teff, logg, abun, ebv,
vgamma, Rv. Grids of LD coeffs dedicated to the targets can now easily be computed.
These grid computations compute LD coefficients from the specific intensities
(thanks to Steven for the core functionality).
(6) Removed the ability to interpolate full SEDs of specificic intensities,
we only need them to create the LD coeff grids anyway. This simplifies the code
considerably
(7) Added reddening module with different reddening laws and conversions between
extinction passbands.
(8) Added support for interstellar reddening to the Bodies.
(9) Ported spot drawing to C (cspots.c).
(10) Implemented passband luminosity and third light support and fitting for Bodies
(11) Improved saving/loading DataSets, added a parameter for weighting in overall fitting
(12) Added parameter-per-parameter ASCII representation of Bodies.
(13) Extended constraint framework: common constraints can be added easily now via
the parameters.tools module (or at least can be implemented there). The only
restriction is that the user needs to add functions in the installation
directory. This the cleanest solution that I could come up with the allow
the user to define functions that are persistent over MPI pickling. In any
case, adding these constraints can be considered an advanced feature.
(14) Overall improvement of error messages from Parameter and ParameterSet
(15) Extended interface functions to Parameter and ParameterSet, and improved
their behaviour
(16) Implemented unique_label to Parameter, which is better than using the
"id" function to generate a unique name for the label, since "id" is not
persistent after pickling and unpickling, which is important for incremental
computations of MCMC.
(17) Emcee fitter is now fully operational and allows incremental MCMC runs
(18) Pymc fitter is now fully operational and allows incremental MCMC runs
(19) Implemented comprehensive summary report for MCMC simulations, this will
need to be expanded upon while we accumulate experience with the fitters. For
now, it just displays the mean and std of the trace. There are no quality
checks as of yet, this still needs to be done manually. It also output plots.
(20) Added possibility to accept an MCMC fit: this sets the fitted parameter
values according to the mean of the traces, and recomputes the systems with
those values
(21) Implemented first attempt to automatically plot observations vs models
of a computed system.
(22) Added emcee support for WD.
(23) Disabled log-to-screen when fitting: the logger streams to a file instead
of the screen (only warnings are put on the screen), and switch back to the
screen-logging after fitting is complete. We might add some more options to
tweak this in the future.
(24) Fixed bug in subdivision: the order of the vertices for the middle
triangle was reversed, causing a change in sign of the normal when calling
compute_normals. This didn't show up before because usually the normals are
not recomputed but instead copied from the original triangle.
(25) Ditched pyrotin package, and implemented rotational convolution in pure
Python (with the help of scipy). A fortran dependency just to convolve a
spectrum seemed a little bit overkill (I just ported something to Python!).
(26) Reorganisation of module structure
(27) Removed a lot of obsolete functions
(28) Expanded Gal's setup.py: allow pip install and uninstall of phoebe via the
following commands (the usual build/install is still possible):
    $:> python setup.py sdist
    $:> sudo pip install dist/phoebe-2.0.tar.gz
    $:> sudo pip uninstall phoebe
For this to work properly, I removed the --wd option, but this is probably
not what we want. now wd is always installed
(29) Imported all top-level functions from different modules (i.e. the ones most
(30) Performed a first check for Python 3.x compatibility in a selection of modules
users should often need) to the phoebe namespace.